+++ Firmware/MotorControl/endstop.hpp
@@ -11,6 +11,12 @@ class Endstop {
         bool enabled = false;
         bool is_active_high = false;
         bool pullup = true;
+
+        // custom setters
+        Endstop* parent = nullptr;
+        void set_gpio_num(uint16_t value) { gpio_num = value; parent->update_config(); }
+        void set_enabled(uint32_t value) { enabled = value; parent->update_config(); }
+        void set_debounce_ms(uint32_t value) { debounce_ms = value; parent->update_config(); }
     };
 
     explicit Endstop(Endstop::Config_t& config);
@@ -25,21 +31,6 @@ class Endstop {
     bool get_state();
 
     bool endstop_state_ = false;
-
-    auto make_protocol_definitions() {
-        return make_protocol_member_list(
-            make_protocol_ro_property("endstop_state", &endstop_state_),
-            make_protocol_object("config",
-                make_protocol_property("gpio_num", &config_.gpio_num,
-                                    [](void* ctx) { static_cast<Endstop*>(ctx)->update_config(); }, this),
-                make_protocol_property("enabled", &config_.enabled,
-                                    [](void* ctx) { static_cast<Endstop*>(ctx)->update_config(); }, this),
-                make_protocol_property("offset", &config_.offset),
-                make_protocol_property("is_active_high", &config_.is_active_high),
-                make_protocol_property("pullup", &config_.pullup),
-                make_protocol_property("debounce_ms", &config_.debounce_ms,
-                                    [](void* ctx) { static_cast<Endstop*>(ctx)->update_config(); }, this)));
-    }
 
    private:
     bool pin_state_ = false;