+++ tools/odrive/tests/pwm_input_test.py
@@ -5,7 +5,7 @@ import time
 import math
 import os
 
-from odrive.enums import errors
+from odrive.enums import *
 from test_runner import *
 
 
@@ -81,7 +81,7 @@ class TestPwmInput():
         full_scale = max_val - min_val
         slope, offset, fitted_curve = fit_sawtooth(data, min_val, max_val)
         test_assert_eq(slope, full_scale / 1.0, accuracy=0.001)
-        test_curve_fit(data, fitted_curve, max_mean_err = full_scale * 0.02, inlier_range = full_scale * 0.05, max_outliers = len(data[:,0]) * 0.01)
+        test_curve_fit(data, fitted_curve, max_mean_err = full_scale * 0.05, inlier_range = full_scale * 0.05, max_outliers = len(data[:,0]) * 0.01)